move : log.d.ts 위치 이동
