+++ src/hooks/mutations/log/useLogCreateMutation.ts
@@ -1,7 +1,7 @@
 import { logKeys, searchKeys } from '@/app/actions/keys';
 import { createLog } from '@/app/actions/log-register';
 import { useLogCreationStore } from '@/stores/logCreationStore';
-import { LogFormValues, NewPlace, NewPlaceImage } from '@/types/schema/log';
+import { LogFormValues, NewPlace, NewPlaceImage } from '@/types/log';
 import { uploadPlaces, uploadThumbnail } from '@/utils/upload';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
